Separate out the individual commands Implement matching of features
